first build of our new inhouse stdlib, libdrysoup™

+++ Programs/directed/lib/memcpy.c
@@ -1,11 +1,14 @@
 //does as char's so that its byte by byte
 //credit to gpt
+#include "drysoup.h"
+#include <stddef.h>
 
-void *memcpy(void *dest, const void *src, unsigned int n) {
+//manual definition of memcpy which is called and occasiaonlly by gcc
+void *memcpy(void *dest, const void *src, size_t n){
     char *d = (char *)dest;
     const char *s = (const char *)src;
 
-    for (unsigned int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         d[i] = s[i];
     }
